Fixed bug enemy shots not checked if enemy destroyed. Fixed Person Man spawn on edge of screen. Fixed new wave start before player explode.

+++ game/src/ScoreKeeper.cpp
@@ -41,7 +41,7 @@ void ScoreKeeper::AddToScore(int amount)
 {
 	Score += amount;
 
-	if (Score >= NextBonusScore)
+	if (Score >= (unsigned int)NextBonusScore)
 	{
 		Bonus = true;
 		NextBonusScore += BonusMult;
